Battleship and narwhar changes
Battleship sinly hates the ammo system. The workaroun is a radar that's
as astring as stacionaty, for 2k energy price. Without if the battleship
is blind. Also flak has been readded to naval in the form of narwhal,
which has a weak flak and cna shoot down missiled, but has lost the
radar.